+++ crates/ryo3-jiff/src/ry_iso_week_date.rs
@@ -2,9 +2,8 @@ use crate::errors::map_py_value_err;
 use crate::{JiffWeekday, RyDate};
 use jiff::civil::ISOWeekDate;
 use jiff::Zoned;
-use pyo3::types::{PyAnyMethods, PyType};
-use pyo3::{pyclass, pymethods, Bound, IntoPyObject, PyResult};
-use std::fmt::Display;
+use pyo3::prelude::*;
+use pyo3::types::PyType;
 use std::hash::{DefaultHasher, Hash, Hasher};
 
 #[derive(Debug, Clone)]
