+++ python/ry/ryo3/__init__.pyi
@@ -6,10 +6,30 @@ from os import PathLike
 
 import typing_extensions as te
 
-from ry._types import DateTimeTypedDict, DateTypedDict, TimeSpanTypedDict, TimeTypedDict
-
 from . import http as http
 from ._bytes import Bytes as Bytes
+from ._jiff import Date as Date
+from ._jiff import DateDifference as DateDifference
+from ._jiff import DateTime as DateTime
+from ._jiff import DateTimeDifference as DateTimeDifference
+from ._jiff import DateTimeRound as DateTimeRound
+from ._jiff import ISOWeekDate as ISOWeekDate
+from ._jiff import Offset as Offset
+from ._jiff import SignedDuration as SignedDuration
+from ._jiff import Time as Time
+from ._jiff import TimeDifference as TimeDifference
+from ._jiff import TimeSpan as TimeSpan
+from ._jiff import Timestamp as Timestamp
+from ._jiff import TimestampDifference as TimestampDifference
+from ._jiff import TimestampRound as TimestampRound
+from ._jiff import TimeZone as TimeZone
+from ._jiff import ZonedDateTime as ZonedDateTime
+from ._jiff import ZonedDateTimeDifference as ZonedDateTimeDifference
+from ._jiff import date as date
+from ._jiff import datetime as datetime
+from ._jiff import offset as offset
+from ._jiff import time as time
+from ._jiff import timespan as timespan
 from .http import Headers as Headers
 from .http import HttpStatus as HttpStatus
 from .reqwest import HttpClient as HttpClient
@@ -705,1386 +725,1386 @@ class URL:
     def set_username(self, username: str) -> None: ...
     def socket_addrs(self) -> None: ...
 
-# =============================================================================
-# JIFF
-# =============================================================================
-JIFF_UNIT_STRING = t.Literal[
-    "year",
-    "month",
-    "week",
-    "day",
-    "hour",
-    "minute",
-    "second",
-    "millisecond",
-    "microsecond",
-    "nanosecond",
-]
-
-JIFF_ROUND_MODE_STRING = t.Literal[
-    "ceil",
-    "floor",
-    "expand",
-    "trunc",
-    "half_ceil",
-    "half_floor",
-    "half_expand",
-    "half_trunc",
-    "half_even",
-]
-
-JIFF_WEEKDAY_STRING = t.Literal[
-    "monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"
-]
-
-JIFF_WEEKDAY_INT = t.Literal[
-    1,  # Monday
-    2,  # Tuesday
-    3,  # Wednesday
-    4,  # Thursday
-    5,  # Friday
-    6,  # Saturday
-    7,  # Sunday
-]
-
-class Date:
-    MIN: Date
-    MAX: Date
-    ZERO: Date
-
-    def __init__(self, year: int, month: int, day: int) -> None: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def string(self) -> str: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # PYTHON_CONVERSIONS
-    # =========================================================================
-    def to_pydate(self) -> pydt.date: ...
-    @classmethod
-    def from_pydate(cls: type[Date], date: pydt.date) -> Date: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def year(self) -> int: ...
-    @property
-    def month(self) -> int: ...
-    @property
-    def day(self) -> int: ...
-    @property
-    def weekday(self) -> int: ...
-
-    # =========================================================================
-    # CLASSMETHODS
-    # =========================================================================
-    @classmethod
-    def from_iso_week_date(
-        cls: type[Date], year: int, week: int, weekday: int
-    ) -> Date: ...
-    @classmethod
-    def today(cls: type[Date]) -> Date: ...
-
-    # =========================================================================
-    # STRPTIME/STRFTIME
-    # =========================================================================
-    @classmethod
-    def strptime(cls: type[Date], format: str, string: str) -> Date: ...
-    def strftime(self, format: str) -> str: ...
-
-    # =========================================================================
-    # OPERATORS
-    # =========================================================================
-    def __add__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    @t.overload
-    def __sub__(self, other: Date) -> TimeSpan: ...
-    @t.overload
-    def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    @t.overload
-    def __isub__(self, other: Date) -> TimeSpan: ...
-    @t.overload
-    def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def at(self, hour: int, minute: int, second: int, nanosecond: int) -> DateTime: ...
-    def asdict(self) -> DateTypedDict: ...
-    def astuple(self) -> tuple[int, int, int]: ...
-    def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    def day_of_year(self) -> int: ...
-    def day_of_year_no_leap(self) -> int | None: ...
-    def days_in_month(self) -> int: ...
-    def days_in_year(self) -> int: ...
-    def duration_since(self, other: Date) -> Date: ...
-    def duration_until(self, other: Date) -> Date: ...
-    def era_year(self) -> tuple[int, t.Literal["BCE", "CE"]]: ...
-    def first_of_month(self) -> Date: ...
-    def first_of_year(self) -> Date: ...
-    def iso_week_date(self) -> ISOWeekDate: ...
-    def in_leap_year(self) -> bool: ...
-    def in_tz(self, tz: str) -> ZonedDateTime: ...
-    def intz(self, tz: str) -> ZonedDateTime: ...
-    def last_of_month(self) -> Date: ...
-    def last_of_year(self) -> Date: ...
-    def nth_weekday(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> Date: ...
-    def nth_weekday_of_month(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> Date: ...
-    def saturating_add(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    def saturating_sub(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-    def series(self, span: TimeSpan) -> t.Iterator[Date]: ...
-    def to_datetime(self, time: Time) -> DateTime: ...
-    def tomorrow(self) -> Date: ...
-    def yesterday(self) -> Date: ...
-
-    # =========================================================================
-    # SINCE/UNTIL
-    # =========================================================================
-    def _since(self, other: DateDifference) -> TimeSpan: ...
-    def _until(self, other: DateDifference) -> TimeSpan: ...
-    def since(
-        self,
-        other: Date | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def until(
-        self,
-        other: Date | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-
-    # =========================================================================
-    # INSTANCE METHODS W/ OVERLOADS
-    # =========================================================================
-    @t.overload
-    def checked_sub(self, other: Date) -> TimeSpan: ...
-    @t.overload
-    def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
-
-class DateDifference:
-    def __init__(
-        self,
-        date: Date,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def smallest(self, unit: JIFF_UNIT_STRING) -> DateDifference: ...
-    def largest(self, unit: JIFF_UNIT_STRING) -> DateDifference: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateDifference: ...
-    def increment(self, increment: int) -> DateDifference: ...
-
-class Time:
-    MIN: Time
-    MAX: Time
-
-    def __init__(
-        self,
-        hour: int = 0,
-        minute: int = 0,
-        second: int = 0,
-        nanosecond: int = 0,
-    ) -> None: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def string(self) -> str: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-    def __add__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    @t.overload
-    def __sub__(self, other: Time) -> TimeSpan: ...
-    @t.overload
-    def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    @t.overload
-    def __isub__(self, other: Time) -> TimeSpan: ...
-    @t.overload
-    def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-
-    # =========================================================================
-    # STRPTIME/STRFTIME/PARSE
-    # =========================================================================
-    @classmethod
-    def strptime(cls: type[Time], format: str, string: str) -> Time: ...
-    def strftime(self, format: str) -> str: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    def to_pytime(self) -> pydt.time: ...
-    @classmethod
-    def from_pytime(cls: type[Time], t: pydt.time) -> Time: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def midnight(cls: type[Time]) -> Time: ...
-    @classmethod
-    def now(cls: type[Time]) -> Time: ...
-    @classmethod
-    def parse(cls: type[Time], s: str) -> Time: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def hour(self) -> int: ...
-    @property
-    def minute(self) -> int: ...
-    @property
-    def second(self) -> int: ...
-    @property
-    def millisecond(self) -> int: ...
-    @property
-    def microsecond(self) -> int: ...
-    @property
-    def nanosecond(self) -> int: ...
-    @property
-    def subsec_nanosecond(self) -> None: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def astuple(self) -> tuple[int, int, int, int]: ...
-    def asdict(self) -> TimeTypedDict: ...
-    def series(self, span: TimeSpan) -> t.Iterator[Time]: ...
-    def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    @t.overload
-    def checked_sub(self, other: Time) -> TimeSpan: ...
-    @t.overload
-    def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def saturating_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def saturating_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def wrapping_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def wrapping_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
-    def on(self, year: int, month: int, day: int) -> DateTime: ...
-    def duration_until(self, other: Time) -> SignedDuration: ...
-    def duration_since(self, other: Time) -> SignedDuration: ...
-    def round(
-        self,
-        smallest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> Time: ...
-    def to_datetime(self, d: Date) -> DateTime: ...
-
-    # =========================================================================
-    # SINCE/UNTIL
-    # =========================================================================
-    def _since(self, other: TimeDifference) -> TimeSpan: ...
-    def _until(self, other: TimeDifference) -> TimeSpan: ...
-    def since(
-        self,
-        other: Time | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def until(
-        self,
-        other: Time | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-
-class TimeDifference:
-    def __init__(
-        self,
-        date: Time,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def smallest(self, unit: JIFF_UNIT_STRING) -> TimeDifference: ...
-    def largest(self, unit: JIFF_UNIT_STRING) -> TimeDifference: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimeDifference: ...
-    def increment(self, increment: int) -> TimeDifference: ...
-
-class DateTime:
-    MIN: DateTime
-    MAX: DateTime
-    ZERO: DateTime
-
-    def __init__(
-        self,
-        year: int,
-        month: int,
-        day: int,
-        hour: int = 0,
-        minute: int = 0,
-        second: int = 0,
-        nanosecond: int = 0,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def string(self) -> str: ...
-
-    # =========================================================================
-    # STRPTIME/STRFTIME/PARSE
-    # =========================================================================
-    def strftime(self, format: str) -> str: ...
-    @classmethod
-    def strptime(cls: type[DateTime], format: str, string: str) -> DateTime: ...
-    @classmethod
-    def parse(cls: type[DateTime], s: str) -> DateTime: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    @classmethod
-    def from_pydatetime(cls: type[DateTime], dt: pydt.datetime) -> DateTime: ...
-    def to_pydatetime(self) -> pydt.datetime: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def now(cls: type[DateTime]) -> DateTime: ...
-    @classmethod
-    def from_parts(cls: type[DateTime], date: Date, time: Time) -> DateTime: ...
-
-    # =========================================================================
-    # OPERATORS
-    # =========================================================================
-    def __repr__(self) -> str: ...
-    def __add__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-    def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-    @t.overload
-    def __sub__(self, other: DateTime) -> TimeSpan: ...
-    @t.overload
-    def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-    @t.overload
-    def __isub__(self, other: DateTime) -> TimeSpan: ...
-    @t.overload
-    def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-    def saturating_add(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> DateTime: ...
-    def in_tz(self, tz: str) -> ZonedDateTime: ...
-    def intz(self, tz: str) -> ZonedDateTime: ...
-    def iso_week_date(self) -> tuple[int, int, int]: ...
-    def date(self) -> Date: ...
-    def time(self) -> Time: ...
-    def series(self, span: TimeSpan) -> t.Iterator[DateTime]: ...
-    def asdict(self) -> DateTimeTypedDict: ...
-    def round(self, options: JIFF_UNIT_STRING | DateTimeRound) -> DateTime: ...
-    def yesterday(self) -> DateTime: ...
-
-    # =========================================================================
-    # INSTANCE METHODS W/ OVERLOADS
-    # =========================================================================
-    @t.overload
-    def checked_sub(self, other: DateTime) -> TimeSpan: ...
-    @t.overload
-    def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
-    @t.overload
-    def saturating_sub(self, other: DateTime) -> TimeSpan: ...
-    @t.overload
-    def saturating_sub(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> DateTime: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def year(self) -> int: ...
-    @property
-    def month(self) -> int: ...
-    @property
-    def day(self) -> int: ...
-    @property
-    def hour(self) -> int: ...
-    @property
-    def minute(self) -> int: ...
-    @property
-    def second(self) -> int: ...
-    @property
-    def millisecond(self) -> int: ...
-    @property
-    def microsecond(self) -> int: ...
-    @property
-    def nanosecond(self) -> int: ...
-    @property
-    def subsec_nanosecond(self) -> int: ...
-    @property
-    def weekday(self) -> int: ...
-
-    # =========================================================================
-    # NOT IMPLEMENTED & NOT TYPED
-    # =========================================================================
-    def first_of_month(self) -> DateTime: ...
-    def first_of_year(self) -> DateTime: ...
-    def day_of_year(self) -> int: ...
-    def day_of_year_no_leap(self) -> int | None: ...
-    def days_in_month(self) -> int: ...
-    def days_in_year(self) -> int: ...
-    def era_year(self) -> tuple[int, t.Literal["BCE", "CE"]]: ...
-    def duration_since(self, other: DateTime) -> SignedDuration: ...
-    def duration_until(self, other: DateTime) -> SignedDuration: ...
-    def in_leap_year(self) -> bool: ...
-    def end_of_day(self) -> DateTime: ...
-    def nth_weekday(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> DateTime: ...
-    def nth_weekday_of_month(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> DateTime: ...
-    def last_of_month(self) -> DateTime: ...
-    def last_of_year(self) -> DateTime: ...
-    def start_of_day(self) -> DateTime: ...
-    def tomorrow(self) -> DateTime: ...
-    def to_zoned(self, tz: TimeZone) -> ZonedDateTime: ...
-
-    # =========================================================================
-    # SINCE/UNTIL
-    # =========================================================================
-    def _since(self, other: DateTimeDifference) -> TimeSpan: ...
-    def _until(self, other: DateTimeDifference) -> TimeSpan: ...
-    def since(
-        self,
-        other: Date | Time | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def until(
-        self,
-        other: Date | Time | DateTime | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-
-class DateTimeDifference:
-    def __init__(
-        self,
-        date: DateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def smallest(self, unit: JIFF_UNIT_STRING) -> DateTimeDifference: ...
-    def largest(self, unit: JIFF_UNIT_STRING) -> DateTimeDifference: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateTimeDifference: ...
-    def increment(self, increment: int) -> DateTimeDifference: ...
-
-class TimeZone:
-    def __init__(self, name: str) -> None: ...
-    def __eq__(self, other: object) -> bool: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    def to_pytzinfo(self) -> pydt.tzinfo: ...
-    @classmethod
-    def from_pytzinfo(cls: type[TimeZone], tz: pydt.tzinfo) -> TimeZone: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def name(self) -> str: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def fixed(cls: type[TimeZone], offset: Offset) -> TimeZone: ...
-    @classmethod
-    def get(cls: type[TimeZone], name: str) -> TimeZone: ...
-    @classmethod
-    def posix(cls: type[TimeZone], name: str) -> TimeZone: ...
-    @classmethod
-    def system(cls: type[TimeZone]) -> TimeZone: ...
-    @classmethod
-    def try_system(cls: type[TimeZone]) -> TimeZone: ...
-    @classmethod
-    def tzif(cls: type[TimeZone], name: str, data: bytes) -> TimeZone: ...
-    @classmethod
-    def utc(cls: type[TimeZone]) -> TimeZone: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def iana_name(self) -> str | None: ...
-    def to_datetime(self, dt: Timestamp) -> DateTime: ...
-    def to_offset(self, timestamp: Timestamp) -> tuple[Offset, bool, str]: ...
-    def to_timestamp(self, dt: DateTime) -> Timestamp: ...
-    def to_zoned(self, other: DateTime) -> ZonedDateTime: ...
-
-    # =========================================================================
-    # NOT IMPLEMENTED
-    # =========================================================================
-    def to_ambiguous_timestamp(self) -> t.NoReturn: ...
-    def to_ambiguous_zoned(self) -> t.NoReturn: ...
-
-class SignedDuration:
-    MIN: SignedDuration
-    MAX: SignedDuration
-    ZERO: SignedDuration
-
-    def __init__(self, secs: int = 0, nanos: int = 0) -> None: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-    def __hash__(self) -> int: ...
-    def __mul__(self, other: int) -> SignedDuration: ...
-    def __rmul__(self, other: int) -> SignedDuration: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: object) -> bool: ...
-    def __le__(self, other: object) -> bool: ...
-    def __gt__(self, other: object) -> bool: ...
-    def __ge__(self, other: object) -> bool: ...
-    def __neg__(self) -> SignedDuration: ...
-    def __add__(self, other: SignedDuration) -> SignedDuration: ...
-    def __abs__(self) -> SignedDuration: ...
-    def __div__(self, other: int) -> SignedDuration: ...
-    def abs(self) -> SignedDuration: ...
-    def unsigned_abs(self) -> Duration: ...
-    def __richcmp__(self, other: SignedDuration | pydt.timedelta, op: int) -> bool: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def __str__(self) -> str: ...
-    def string(self, human: bool = False) -> str: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    @classmethod
-    def from_pytimedelta(
-        cls: type[SignedDuration], td: pydt.timedelta
-    ) -> SignedDuration: ...
-    def to_pytimedelta(self) -> pydt.timedelta: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def parse(cls: type[SignedDuration], s: str) -> SignedDuration: ...
-    @classmethod
-    def from_hours(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_micros(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_millis(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_mins(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_nanos(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_secs(cls: type[SignedDuration], n: int) -> SignedDuration: ...
-    @classmethod
-    def from_secs_f32(cls: type[SignedDuration], n: float) -> SignedDuration: ...
-    @classmethod
-    def from_secs_f64(cls: type[SignedDuration], n: float) -> SignedDuration: ...
-    @classmethod
-    def try_from_secs_f32(cls: type[SignedDuration], n: float) -> SignedDuration: ...
-    @classmethod
-    def try_from_secs_f64(cls: type[SignedDuration], n: float) -> SignedDuration: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def is_negative(self) -> bool: ...
-    @property
-    def is_zero(self) -> bool: ...
-    @property
-    def secs(self) -> int: ...
-    @property
-    def nanos(self) -> int: ...
-    @property
-    def days(self) -> int: ...
-    @property
-    def seconds(self) -> int: ...
-    @property
-    def microseconds(self) -> int: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def as_hours(self) -> int: ...
-    def as_micros(self) -> int: ...
-    def as_millis(self) -> int: ...
-    def as_millis_f32(self) -> float: ...
-    def as_millis_f64(self) -> float: ...
-    def as_mins(self) -> int: ...
-    def as_nanos(self) -> int: ...
-    def as_secs(self) -> int: ...
-    def as_secs_f32(self) -> float: ...
-    def as_secs_f64(self) -> float: ...
-    def checked_add(self, other: SignedDuration) -> SignedDuration | None: ...
-    def checked_div(self, other: int) -> SignedDuration | None: ...
-    def checked_mul(self, other: int) -> SignedDuration | None: ...
-    def checked_neg(self) -> SignedDuration | None: ...
-    def checked_sub(self, other: SignedDuration) -> SignedDuration | None: ...
-    def div_duration_f32(self, other: SignedDuration) -> float: ...
-    def div_duration_f64(self, other: SignedDuration) -> float: ...
-    def div_f32(self, other: int) -> float: ...
-    def div_f64(self, other: int) -> float: ...
-    def is_positive(self) -> bool: ...
-    def mul_f32(self, other: int) -> SignedDuration: ...
-    def mul_f64(self, other: int) -> SignedDuration: ...
-    def saturating_add(self, other: SignedDuration) -> SignedDuration: ...
-    def saturating_mul(self, other: int) -> SignedDuration: ...
-    def saturating_sub(self, other: SignedDuration) -> SignedDuration: ...
-    def signum(self) -> t.Literal[-1, 0, 1]: ...
-    def subsec_micros(self) -> int: ...
-    def subsec_millis(self) -> int: ...
-    def subsec_nanos(self) -> int: ...
-    def to_timespan(self) -> TimeSpan: ...
-
-# put in quotes to avoid ruff F821 - undefined name
-_TimeSpanArithmeticSingle = TimeSpan | Duration | SignedDuration
-_TimeSpanArithmeticTuple = tuple[
-    _TimeSpanArithmeticSingle, ZonedDateTime | Date | DateTime
-]
-TimeSpanArithmetic = _TimeSpanArithmeticSingle | _TimeSpanArithmeticTuple
-
-class TimeSpan:
-    def __init__(
-        self,
-        years: int = 0,
-        months: int = 0,
-        weeks: int = 0,
-        days: int = 0,
-        hours: int = 0,
-        minutes: int = 0,
-        seconds: int = 0,
-        milliseconds: int = 0,
-        microseconds: int = 0,
-        nanoseconds: int = 0,
-    ) -> None: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def string(self, human: bool = False) -> str: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def repr_full(self) -> str: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    @classmethod
-    def from_pytimedelta(cls, td: pydt.timedelta) -> TimeSpan: ...
-    def to_pytimedelta(self) -> pydt.timedelta: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def parse(cls, s: str) -> TimeSpan: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def is_positive(self) -> bool: ...
-    @property
-    def is_negative(self) -> bool: ...
-    @property
-    def is_zero(self) -> bool: ...
-    @property
-    def years(self) -> int: ...
-    @property
-    def months(self) -> int: ...
-    @property
-    def weeks(self) -> int: ...
-    @property
-    def days(self) -> int: ...
-    @property
-    def hours(self) -> int: ...
-    @property
-    def minutes(self) -> int: ...
-    @property
-    def seconds(self) -> int: ...
-    @property
-    def milliseconds(self) -> int: ...
-    @property
-    def microseconds(self) -> int: ...
-    @property
-    def nanoseconds(self) -> int: ...
-
-    # =========================================================================
-    # OPERATORS
-    # =========================================================================
-    def __add__(
-        self,
-        val: TimeSpanArithmetic,
-    ) -> te.Self: ...
-    def __sub__(
-        self,
-        val: TimeSpanArithmetic,
-    ) -> te.Self: ...
-    def __mul__(self, other: int) -> te.Self: ...
-    def __neg__(self) -> te.Self: ...
-    def __abs__(self) -> te.Self: ...
-    def __invert__(self) -> te.Self: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ge__(self, other: TimeSpan) -> bool: ...
-    def __gt__(self, other: TimeSpan) -> bool: ...
-    def __le__(self, other: TimeSpan) -> bool: ...
-    def __lt__(self, other: TimeSpan) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __rmul__(self, other: TimeSpan) -> bool: ...
-    def __hash__(self) -> int: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-
-    def abs(self) -> te.Self: ...
-    def asdict(self) -> TimeSpanTypedDict: ...
-    def checked_add(self, val: TimeSpanArithmetic) -> te.Self: ...
-    def checked_mul(self, other: int) -> te.Self: ...
-    def checked_sub(self, val: TimeSpanArithmetic) -> te.Self: ...
-    def compare(
-        self, other: TimeSpan, relative: ZonedDateTime | DateTime | Date | None = None
-    ) -> int: ...
-    def negate(self) -> te.Self: ...
-    def replace(
-        self,
-        years: int | None = None,
-        months: int | None = None,
-        weeks: int | None = None,
-        days: int | None = None,
-        hours: int | None = None,
-        minutes: int | None = None,
-        seconds: int | None = None,
-        milliseconds: int | None = None,
-        microseconds: int | None = None,
-        nanoseconds: int | None = None,
-    ) -> te.Self: ...
-    def round(
-        self,
-        smallest: JIFF_UNIT_STRING,
-        increment: int = 1,
-        *,
-        relative: ZonedDateTime | Date | DateTime | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-    ) -> te.Self: ...
-    def signum(self) -> t.Literal[-1, 0, 1]: ...
-    def to_signed_duration(
-        self, relative: ZonedDateTime | Date | DateTime
-    ) -> SignedDuration: ...
-    def total(
-        self,
-        unit: JIFF_UNIT_STRING,
-        relative: ZonedDateTime | Date | DateTime | None = None,
-    ) -> int: ...
-    def total_seconds(self) -> int: ...
-    def try_years(self, years: int) -> te.Self: ...
-    def try_months(self, months: int) -> te.Self: ...
-    def try_weeks(self, weeks: int) -> te.Self: ...
-    def try_days(self, days: int) -> te.Self: ...
-    def try_hours(self, hours: int) -> te.Self: ...
-    def try_minutes(self, minutes: int) -> te.Self: ...
-    def try_seconds(self, seconds: int) -> te.Self: ...
-    def try_milliseconds(self, milliseconds: int) -> te.Self: ...
-    def try_microseconds(self, microseconds: int) -> te.Self: ...
-    def try_nanoseconds(self, nanoseconds: int) -> te.Self: ...
-
-    # -------------------------------------------------------------------------
-    # PANIC-INDUCING METHODS
-    # -------------------------------------------------------------------------
-    def _years(self, years: int) -> te.Self: ...
-    def _months(self, months: int) -> te.Self: ...
-    def _weeks(self, weeks: int) -> te.Self: ...
-    def _days(self, days: int) -> te.Self: ...
-    def _hours(self, hours: int) -> te.Self: ...
-    def _minutes(self, minutes: int) -> te.Self: ...
-    def _seconds(self, seconds: int) -> te.Self: ...
-    def _milliseconds(self, milliseconds: int) -> te.Self: ...
-    def _microseconds(self, microseconds: int) -> te.Self: ...
-    def _nanoseconds(self, nanoseconds: int) -> te.Self: ...
-
-class Timestamp:
-    """
-    A representation of a timestamp with second and nanosecond precision.
-    """
-
-    MIN: Timestamp
-    MAX: Timestamp
-    UNIX_EPOCH: Timestamp
-
-    def __init__(
-        self, second: int | None = None, nanosecond: int | None = None
-    ) -> None: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def now(cls) -> Timestamp: ...
-    @classmethod
-    def parse(cls, s: str) -> Timestamp: ...
-    @classmethod
-    def from_millisecond(cls, millisecond: int) -> Timestamp: ...
-    @classmethod
-    def from_microsecond(cls, microsecond: int) -> Timestamp: ...
-    @classmethod
-    def from_nanosecond(cls, nanosecond: int) -> Timestamp: ...
-    @classmethod
-    def from_second(cls, second: int) -> Timestamp: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-    def __add__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ge__(self, other: Timestamp) -> bool: ...
-    def __gt__(self, other: Timestamp) -> bool: ...
-    def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    def __le__(self, other: Timestamp) -> bool: ...
-    def __lt__(self, other: Timestamp) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __richcmp__(self, other: Timestamp, op: int) -> bool: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS W/ OVERLOADS
-    # =========================================================================
-    @t.overload
-    def __isub__(self, other: Timestamp) -> TimeSpan: ...
-    @t.overload
-    def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    @t.overload
-    def __sub__(self, other: Timestamp) -> TimeSpan: ...
-    @t.overload
-    def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-
-    # =========================================================================
-    # STRPTIME/STRFTIME
-    # =========================================================================
-    def strftime(self, format: str) -> str: ...
-    @classmethod
-    def strptime(cls, format: str, input: str) -> Timestamp: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-
-    def as_microsecond(self) -> int: ...
-    def as_millisecond(self) -> int: ...
-    def as_nanosecond(self) -> int: ...
-    def as_second(self) -> int: ...
-    def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Timestamp: ...
-    @t.overload
-    def checked_sub(self, other: Timestamp) -> TimeSpan: ...
-    @t.overload
-    def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Timestamp: ...
-    def display_with_offset(self, offset: Offset) -> str: ...
-    def in_tz(self, tz: str) -> ZonedDateTime: ...
-    def intz(self, tz: str) -> ZonedDateTime:
-        """Deprecated ~ use `in_tz`"""
-
-    def is_zero(self) -> bool: ...
-    def saturating_add(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> Timestamp: ...
-    def saturating_sub(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> Timestamp: ...
-    def series(self, span: TimeSpan) -> t.Iterator[Timestamp]: ...
-    def signum(self) -> t.Literal[-1, 0, 1]: ...
-    def string(self) -> str: ...
-    def subsec_microsecond(self) -> int: ...
-    def subsec_millisecond(self) -> int: ...
-    def subsec_nanosecond(self) -> int: ...
-    def to_zoned(self, time_zone: TimeZone) -> ZonedDateTime: ...
-
-    # =========================================================================
-    # SINCE/UNTIL
-    # =========================================================================
-    def _since(self, other: TimestampDifference) -> TimeSpan: ...
-    def _until(self, other: TimestampDifference) -> TimeSpan: ...
-    def since(
-        self,
-        other: Timestamp | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def until(
-        self,
-        other: Timestamp | ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def duration_since(self, other: Timestamp) -> SignedDuration: ...
-    def duration_until(self, other: Timestamp) -> SignedDuration: ...
-    def round(
-        self,
-        unit: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> Timestamp: ...
-    def _round(self, options: TimestampRound) -> Timestamp: ...
-
-class TimestampDifference:
-    def __init__(
-        self,
-        date: Timestamp,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def smallest(self, unit: JIFF_UNIT_STRING) -> TimestampDifference: ...
-    def largest(self, unit: JIFF_UNIT_STRING) -> TimestampDifference: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimestampDifference: ...
-    def increment(self, increment: int) -> TimestampDifference: ...
-
-class ZonedDateTime:
-    def __init__(self, timestamp: Timestamp, time_zone: TimeZone) -> None: ...
-
-    # =========================================================================
-    # PYTHON CONVERSIONS
-    # =========================================================================
-    @classmethod
-    def from_pydatetime(
-        cls: type[ZonedDateTime], dt: pydt.datetime
-    ) -> ZonedDateTime: ...
-    def to_pydatetime(self) -> pydt.datetime: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def now(cls: type[ZonedDateTime], tz: str | None = None) -> ZonedDateTime: ...
-    @classmethod
-    def utcnow(cls: type[ZonedDateTime]) -> ZonedDateTime: ...
-    @classmethod
-    def parse(cls: type[ZonedDateTime], s: str) -> ZonedDateTime: ...
-    @classmethod
-    def from_rfc2822(cls: type[ZonedDateTime], s: str) -> ZonedDateTime: ...
-
-    # =========================================================================
-    # STRPTIME/STRFTIME
-    # =========================================================================
-    @classmethod
-    def strptime(
-        cls: type[ZonedDateTime], format: str, input: str
-    ) -> ZonedDateTime: ...
-    def strftime(self, format: str) -> str: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def year(self) -> int: ...
-    @property
-    def month(self) -> int: ...
-    @property
-    def day(self) -> int: ...
-    @property
-    def hour(self) -> int: ...
-    @property
-    def minute(self) -> int: ...
-    @property
-    def second(self) -> int: ...
-    @property
-    def millisecond(self) -> int: ...
-    @property
-    def microsecond(self) -> int: ...
-    @property
-    def nanosecond(self) -> int: ...
-    @property
-    def subsec_nanosecond(self) -> int: ...
-    @property
-    def weekday(self) -> int: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def __str__(self) -> str: ...
-    def string(self) -> str: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-    def __add__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ge__(self, other: ZonedDateTime) -> bool: ...
-    def __gt__(self, other: ZonedDateTime) -> bool: ...
-    def __hash__(self) -> int: ...
-    def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    def __le__(self, other: ZonedDateTime) -> bool: ...
-    def __lt__(self, other: ZonedDateTime) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __richcmp__(self, other: ZonedDateTime, op: int) -> bool: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS W/ OVERLOADS
-    # =========================================================================
-    @t.overload
-    def __isub__(self, other: ZonedDateTime) -> TimeSpan: ...
-    @t.overload
-    def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    @t.overload
-    def __sub__(self, other: ZonedDateTime) -> TimeSpan: ...
-    @t.overload
-    def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def astimezone(self, tz: str) -> ZonedDateTime: ...
-    def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    @t.overload
-    def checked_sub(self, other: ZonedDateTime) -> TimeSpan: ...
-    @t.overload
-    def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
-    def date(self) -> Date: ...
-    def datetime(self) -> DateTime: ...
-    def iso_week_date(self) -> ISOWeekDate: ...
-    def day_of_year(self) -> int: ...
-    def day_of_year_no_leap(self) -> int | None: ...
-    def days_in_month(self) -> int: ...
-    def days_in_year(self) -> int: ...
-    def duration_since(self, other: ZonedDateTime) -> SignedDuration: ...
-    def duration_until(self, other: ZonedDateTime) -> SignedDuration: ...
-    def end_of_day(self) -> ZonedDateTime: ...
-    def era_year(self) -> tuple[int, t.Literal["CE", "BCE"]]: ...
-    def first_of_month(self) -> ZonedDateTime: ...
-    def first_of_year(self) -> ZonedDateTime: ...
-    def in_leap_year(self) -> bool: ...
-    def in_tz(self, tz: str) -> te.Self: ...
-    def intz(self, tz: str) -> te.Self: ...
-    def inutc(self) -> ZonedDateTime: ...
-    def last_of_month(self) -> ZonedDateTime: ...
-    def last_of_year(self) -> ZonedDateTime: ...
-    def nth_weekday(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> Date: ...
-    def nth_weekday_of_month(
-        self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
-    ) -> Date: ...
-    def offset(self) -> Offset: ...
-    def round(self, options: JIFF_UNIT_STRING | DateTimeRound) -> te.Self: ...
-    def saturating_add(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> te.Self: ...
-    @t.overload
-    def saturating_sub(self, other: ZonedDateTime) -> TimeSpan: ...
-    @t.overload
-    def saturating_sub(
-        self, other: TimeSpan | SignedDuration | Duration
-    ) -> te.Self: ...
-    def start_of_day(self) -> ZonedDateTime: ...
-    def time(self) -> Time: ...
-    def time_zone(self) -> TimeZone: ...
-    def timestamp(self) -> Timestamp: ...
-    def timezone(self) -> TimeZone: ...
-    def to_rfc2822(self) -> str: ...
-    def tomorrow(self) -> ZonedDateTime: ...
-    def with_time_zone(self, tz: TimeZone) -> ZonedDateTime: ...
-    def yesterday(self) -> ZonedDateTime: ...
-
-    # =========================================================================
-    # SINCE/UNTIL
-    # =========================================================================
-    def since(
-        self,
-        other: ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-    def until(
-        self,
-        other: ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> TimeSpan: ...
-
-class ZonedDateTimeDifference:
-    def __init__(
-        self,
-        date: ZonedDateTime,
-        *,
-        smallest: JIFF_UNIT_STRING | None = None,
-        largest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int | None = None,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def smallest(self, unit: JIFF_UNIT_STRING) -> ZonedDateTimeDifference: ...
-    def largest(self, unit: JIFF_UNIT_STRING) -> ZonedDateTimeDifference: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> ZonedDateTimeDifference: ...
-    def increment(self, increment: int) -> ZonedDateTimeDifference: ...
-
-class ISOWeekDate:
-    MIN: ISOWeekDate
-    MAX: ISOWeekDate
-    ZERO: ISOWeekDate
-
-    def __init__(
-        self, year: int, week: int, weekday: JIFF_WEEKDAY_INT | JIFF_WEEKDAY_STRING
-    ) -> None: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: ISOWeekDate) -> bool: ...
-    def __le__(self, other: ISOWeekDate) -> bool: ...
-    def __gt__(self, other: ISOWeekDate) -> bool: ...
-    def __ge__(self, other: ISOWeekDate) -> bool: ...
-    def __hash__(self) -> int: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # CLASS METHODS
-    # =========================================================================
-    @classmethod
-    def from_date(cls: type[ISOWeekDate], date: Date) -> ISOWeekDate: ...
-    @classmethod
-    def today(cls: type[ISOWeekDate]) -> ISOWeekDate: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def year(self) -> int: ...
-    @property
-    def week(self) -> int: ...
-    @property
-    def weekday(self) -> JIFF_WEEKDAY_INT: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def date(self) -> Date: ...
-
-class TimestampRound:
-    def __init__(
-        self,
-        smallest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int = 1,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def __eq__(self, other: object) -> bool: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimestampRound: ...
-    def smallest(self, smallest: JIFF_UNIT_STRING) -> TimestampRound: ...
-    def increment(self, increment: int) -> TimestampRound: ...
-    def _smallest(self) -> JIFF_UNIT_STRING: ...
-    def _mode(self) -> JIFF_ROUND_MODE_STRING: ...
-    def _increment(self) -> int: ...
-    def replace(
-        self,
-        smallest: JIFF_UNIT_STRING | None,
-        mode: JIFF_ROUND_MODE_STRING | None,
-        increment: int | None,
-    ) -> TimestampRound: ...
-
-class DateTimeRound:
-    def __init__(
-        self,
-        smallest: JIFF_UNIT_STRING | None = None,
-        mode: JIFF_ROUND_MODE_STRING | None = None,
-        increment: int = 1,
-    ) -> None: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-    def __eq__(self, other: object) -> bool: ...
-    def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateTimeRound: ...
-    def smallest(self, smallest: JIFF_UNIT_STRING) -> DateTimeRound: ...
-    def increment(self, increment: int) -> DateTimeRound: ...
-    def _smallest(self) -> JIFF_UNIT_STRING: ...
-    def _mode(self) -> JIFF_ROUND_MODE_STRING: ...
-    def _increment(self) -> int: ...
-    def replace(
-        self,
-        smallest: JIFF_UNIT_STRING | None,
-        mode: JIFF_ROUND_MODE_STRING | None,
-        increment: int | None,
-    ) -> DateTimeRound: ...
-
-class Offset:
-    MIN: Offset
-    MAX: Offset
-    UTC: Offset
-    ZERO: Offset
-
-    def __init__(
-        self, hours: int | None = None, seconds: int | None = None
-    ) -> None: ...
-
-    # =========================================================================
-    # STRING
-    # =========================================================================
-    def string(self) -> str: ...
-    def __str__(self) -> str: ...
-    def __repr__(self) -> str: ...
-
-    # =========================================================================
-    # OPERATORS/DUNDERS
-    # =========================================================================
-    def __neg__(self) -> Offset: ...
-    def __eq__(self, other: object) -> bool: ...
-    def __ne__(self, other: object) -> bool: ...
-    def __lt__(self, other: Offset) -> bool: ...
-    def __le__(self, other: Offset) -> bool: ...
-    def __gt__(self, other: Offset) -> bool: ...
-    def __ge__(self, other: Offset) -> bool: ...
-    def __hash__(self) -> int: ...
-
-    # =========================================================================
-    # PROPERTIES
-    # =========================================================================
-    @property
-    def seconds(self) -> int: ...
-    @property
-    def is_negative(self) -> bool: ...
-    @property
-    def is_positive(self) -> bool: ...
-
-    # =========================================================================
-    # FROM
-    # =========================================================================
-    @classmethod
-    def utc(cls: type[Offset]) -> Offset: ...
-    @classmethod
-    def from_hours(cls: type[Offset], hours: int) -> Offset: ...
-    @classmethod
-    def from_seconds(cls: type[Offset], seconds: int) -> Offset: ...
-
-    # =========================================================================
-    # TO
-    # =========================================================================
-    def to_datetime(self, timestamp: Timestamp) -> DateTime: ...
-    def to_timestamp(self, datetime: DateTime) -> Timestamp: ...
-    def to_timezone(self) -> TimeZone: ...
-
-    # =========================================================================
-    # INSTANCE METHODS
-    # =========================================================================
-    def checked_add(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
-    def checked_sub(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
-    def duration_since(self, other: Offset) -> SignedDuration: ...
-    def duration_until(self, other: Offset) -> SignedDuration: ...
-    def negate(self) -> Offset: ...
-    def saturating_add(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
-    def saturating_sub(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
-    def since(self, other: Offset) -> TimeSpan: ...
-    def until(self, other: Offset) -> TimeSpan: ...
-
-def date(year: int, month: int, day: int) -> Date: ...
-def time(
-    hour: int = 0, minute: int = 0, second: int = 0, nanosecond: int = 0
-) -> Time: ...
-def datetime(
-    year: int,
-    month: int,
-    day: int,
-    hour: int = 0,
-    minute: int = 0,
-    second: int = 0,
-    nanosecond: int = 0,
-) -> DateTime: ...
-def timespan(
-    *,
-    years: int = 0,
-    months: int = 0,
-    weeks: int = 0,
-    days: int = 0,
-    hours: int = 0,
-    minutes: int = 0,
-    seconds: int = 0,
-    milliseconds: int = 0,
-    microseconds: int = 0,
-    nanoseconds: int = 0,
-    unchecked: bool = False,
-) -> TimeSpan: ...
-def offset(hours: int) -> Offset: ...
+# # =============================================================================
+# # JIFF
+# # =============================================================================
+# JIFF_UNIT_STRING = t.Literal[
+#     "year",
+#     "month",
+#     "week",
+#     "day",
+#     "hour",
+#     "minute",
+#     "second",
+#     "millisecond",
+#     "microsecond",
+#     "nanosecond",
+# ]
+#
+# JIFF_ROUND_MODE_STRING = t.Literal[
+#     "ceil",
+#     "floor",
+#     "expand",
+#     "trunc",
+#     "half_ceil",
+#     "half_floor",
+#     "half_expand",
+#     "half_trunc",
+#     "half_even",
+# ]
+#
+# JIFF_WEEKDAY_STRING = t.Literal[
+#     "monday", "tuesday", "wednesday", "thursday", "friday", "saturday", "sunday"
+# ]
+#
+# JIFF_WEEKDAY_INT = t.Literal[
+#     1,  # Monday
+#     2,  # Tuesday
+#     3,  # Wednesday
+#     4,  # Thursday
+#     5,  # Friday
+#     6,  # Saturday
+#     7,  # Sunday
+# ]
+#
+# class Date:
+#     MIN: Date
+#     MAX: Date
+#     ZERO: Date
+#
+#     def __init__(self, year: int, month: int, day: int) -> None: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def string(self) -> str: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # PYTHON_CONVERSIONS
+#     # =========================================================================
+#     def to_pydate(self) -> pydt.date: ...
+#     @classmethod
+#     def from_pydate(cls: type[Date], date: pydt.date) -> Date: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def year(self) -> int: ...
+#     @property
+#     def month(self) -> int: ...
+#     @property
+#     def day(self) -> int: ...
+#     @property
+#     def weekday(self) -> int: ...
+#
+#     # =========================================================================
+#     # CLASSMETHODS
+#     # =========================================================================
+#     @classmethod
+#     def from_iso_week_date(
+#         cls: type[Date], year: int, week: int, weekday: int
+#     ) -> Date: ...
+#     @classmethod
+#     def today(cls: type[Date]) -> Date: ...
+#
+#     # =========================================================================
+#     # STRPTIME/STRFTIME
+#     # =========================================================================
+#     @classmethod
+#     def strptime(cls: type[Date], format: str, string: str) -> Date: ...
+#     def strftime(self, format: str) -> str: ...
+#
+#     # =========================================================================
+#     # OPERATORS
+#     # =========================================================================
+#     def __add__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     @t.overload
+#     def __sub__(self, other: Date) -> TimeSpan: ...
+#     @t.overload
+#     def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     @t.overload
+#     def __isub__(self, other: Date) -> TimeSpan: ...
+#     @t.overload
+#     def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def at(self, hour: int, minute: int, second: int, nanosecond: int) -> DateTime: ...
+#     def asdict(self) -> DateTypedDict: ...
+#     def astuple(self) -> tuple[int, int, int]: ...
+#     def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     def day_of_year(self) -> int: ...
+#     def day_of_year_no_leap(self) -> int | None: ...
+#     def days_in_month(self) -> int: ...
+#     def days_in_year(self) -> int: ...
+#     def duration_since(self, other: Date) -> Date: ...
+#     def duration_until(self, other: Date) -> Date: ...
+#     def era_year(self) -> tuple[int, t.Literal["BCE", "CE"]]: ...
+#     def first_of_month(self) -> Date: ...
+#     def first_of_year(self) -> Date: ...
+#     def iso_week_date(self) -> ISOWeekDate: ...
+#     def in_leap_year(self) -> bool: ...
+#     def in_tz(self, tz: str) -> ZonedDateTime: ...
+#     def intz(self, tz: str) -> ZonedDateTime: ...
+#     def last_of_month(self) -> Date: ...
+#     def last_of_year(self) -> Date: ...
+#     def nth_weekday(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> Date: ...
+#     def nth_weekday_of_month(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> Date: ...
+#     def saturating_add(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     def saturating_sub(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#     def series(self, span: TimeSpan) -> t.Iterator[Date]: ...
+#     def to_datetime(self, time: Time) -> DateTime: ...
+#     def tomorrow(self) -> Date: ...
+#     def yesterday(self) -> Date: ...
+#
+#     # =========================================================================
+#     # SINCE/UNTIL
+#     # =========================================================================
+#     def _since(self, other: DateDifference) -> TimeSpan: ...
+#     def _until(self, other: DateDifference) -> TimeSpan: ...
+#     def since(
+#         self,
+#         other: Date | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def until(
+#         self,
+#         other: Date | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS W/ OVERLOADS
+#     # =========================================================================
+#     @t.overload
+#     def checked_sub(self, other: Date) -> TimeSpan: ...
+#     @t.overload
+#     def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Date: ...
+#
+# class DateDifference:
+#     def __init__(
+#         self,
+#         date: Date,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def smallest(self, unit: JIFF_UNIT_STRING) -> DateDifference: ...
+#     def largest(self, unit: JIFF_UNIT_STRING) -> DateDifference: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateDifference: ...
+#     def increment(self, increment: int) -> DateDifference: ...
+#
+# class Time:
+#     MIN: Time
+#     MAX: Time
+#
+#     def __init__(
+#         self,
+#         hour: int = 0,
+#         minute: int = 0,
+#         second: int = 0,
+#         nanosecond: int = 0,
+#     ) -> None: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def string(self) -> str: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#     def __add__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     @t.overload
+#     def __sub__(self, other: Time) -> TimeSpan: ...
+#     @t.overload
+#     def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     @t.overload
+#     def __isub__(self, other: Time) -> TimeSpan: ...
+#     @t.overload
+#     def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#
+#     # =========================================================================
+#     # STRPTIME/STRFTIME/PARSE
+#     # =========================================================================
+#     @classmethod
+#     def strptime(cls: type[Time], format: str, string: str) -> Time: ...
+#     def strftime(self, format: str) -> str: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     def to_pytime(self) -> pydt.time: ...
+#     @classmethod
+#     def from_pytime(cls: type[Time], t: pydt.time) -> Time: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def midnight(cls: type[Time]) -> Time: ...
+#     @classmethod
+#     def now(cls: type[Time]) -> Time: ...
+#     @classmethod
+#     def parse(cls: type[Time], s: str) -> Time: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def hour(self) -> int: ...
+#     @property
+#     def minute(self) -> int: ...
+#     @property
+#     def second(self) -> int: ...
+#     @property
+#     def millisecond(self) -> int: ...
+#     @property
+#     def microsecond(self) -> int: ...
+#     @property
+#     def nanosecond(self) -> int: ...
+#     @property
+#     def subsec_nanosecond(self) -> None: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def astuple(self) -> tuple[int, int, int, int]: ...
+#     def asdict(self) -> TimeTypedDict: ...
+#     def series(self, span: TimeSpan) -> t.Iterator[Time]: ...
+#     def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     @t.overload
+#     def checked_sub(self, other: Time) -> TimeSpan: ...
+#     @t.overload
+#     def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def saturating_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def saturating_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def wrapping_add(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def wrapping_sub(self, other: TimeSpan | SignedDuration | Duration) -> Time: ...
+#     def on(self, year: int, month: int, day: int) -> DateTime: ...
+#     def duration_until(self, other: Time) -> SignedDuration: ...
+#     def duration_since(self, other: Time) -> SignedDuration: ...
+#     def round(
+#         self,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> Time: ...
+#     def to_datetime(self, d: Date) -> DateTime: ...
+#
+#     # =========================================================================
+#     # SINCE/UNTIL
+#     # =========================================================================
+#     def _since(self, other: TimeDifference) -> TimeSpan: ...
+#     def _until(self, other: TimeDifference) -> TimeSpan: ...
+#     def since(
+#         self,
+#         other: Time | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def until(
+#         self,
+#         other: Time | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#
+# class TimeDifference:
+#     def __init__(
+#         self,
+#         date: Time,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def smallest(self, unit: JIFF_UNIT_STRING) -> TimeDifference: ...
+#     def largest(self, unit: JIFF_UNIT_STRING) -> TimeDifference: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimeDifference: ...
+#     def increment(self, increment: int) -> TimeDifference: ...
+#
+# class DateTime:
+#     MIN: DateTime
+#     MAX: DateTime
+#     ZERO: DateTime
+#
+#     def __init__(
+#         self,
+#         year: int,
+#         month: int,
+#         day: int,
+#         hour: int = 0,
+#         minute: int = 0,
+#         second: int = 0,
+#         nanosecond: int = 0,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def string(self) -> str: ...
+#
+#     # =========================================================================
+#     # STRPTIME/STRFTIME/PARSE
+#     # =========================================================================
+#     def strftime(self, format: str) -> str: ...
+#     @classmethod
+#     def strptime(cls: type[DateTime], format: str, string: str) -> DateTime: ...
+#     @classmethod
+#     def parse(cls: type[DateTime], s: str) -> DateTime: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     @classmethod
+#     def from_pydatetime(cls: type[DateTime], dt: pydt.datetime) -> DateTime: ...
+#     def to_pydatetime(self) -> pydt.datetime: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def now(cls: type[DateTime]) -> DateTime: ...
+#     @classmethod
+#     def from_parts(cls: type[DateTime], date: Date, time: Time) -> DateTime: ...
+#
+#     # =========================================================================
+#     # OPERATORS
+#     # =========================================================================
+#     def __repr__(self) -> str: ...
+#     def __add__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#     def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#     @t.overload
+#     def __sub__(self, other: DateTime) -> TimeSpan: ...
+#     @t.overload
+#     def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#     @t.overload
+#     def __isub__(self, other: DateTime) -> TimeSpan: ...
+#     @t.overload
+#     def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#     def saturating_add(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> DateTime: ...
+#     def in_tz(self, tz: str) -> ZonedDateTime: ...
+#     def intz(self, tz: str) -> ZonedDateTime: ...
+#     def iso_week_date(self) -> tuple[int, int, int]: ...
+#     def date(self) -> Date: ...
+#     def time(self) -> Time: ...
+#     def series(self, span: TimeSpan) -> t.Iterator[DateTime]: ...
+#     def asdict(self) -> DateTimeTypedDict: ...
+#     def round(self, options: JIFF_UNIT_STRING | DateTimeRound) -> DateTime: ...
+#     def yesterday(self) -> DateTime: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS W/ OVERLOADS
+#     # =========================================================================
+#     @t.overload
+#     def checked_sub(self, other: DateTime) -> TimeSpan: ...
+#     @t.overload
+#     def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> DateTime: ...
+#     @t.overload
+#     def saturating_sub(self, other: DateTime) -> TimeSpan: ...
+#     @t.overload
+#     def saturating_sub(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> DateTime: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def year(self) -> int: ...
+#     @property
+#     def month(self) -> int: ...
+#     @property
+#     def day(self) -> int: ...
+#     @property
+#     def hour(self) -> int: ...
+#     @property
+#     def minute(self) -> int: ...
+#     @property
+#     def second(self) -> int: ...
+#     @property
+#     def millisecond(self) -> int: ...
+#     @property
+#     def microsecond(self) -> int: ...
+#     @property
+#     def nanosecond(self) -> int: ...
+#     @property
+#     def subsec_nanosecond(self) -> int: ...
+#     @property
+#     def weekday(self) -> int: ...
+#
+#     # =========================================================================
+#     # NOT IMPLEMENTED & NOT TYPED
+#     # =========================================================================
+#     def first_of_month(self) -> DateTime: ...
+#     def first_of_year(self) -> DateTime: ...
+#     def day_of_year(self) -> int: ...
+#     def day_of_year_no_leap(self) -> int | None: ...
+#     def days_in_month(self) -> int: ...
+#     def days_in_year(self) -> int: ...
+#     def era_year(self) -> tuple[int, t.Literal["BCE", "CE"]]: ...
+#     def duration_since(self, other: DateTime) -> SignedDuration: ...
+#     def duration_until(self, other: DateTime) -> SignedDuration: ...
+#     def in_leap_year(self) -> bool: ...
+#     def end_of_day(self) -> DateTime: ...
+#     def nth_weekday(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> DateTime: ...
+#     def nth_weekday_of_month(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> DateTime: ...
+#     def last_of_month(self) -> DateTime: ...
+#     def last_of_year(self) -> DateTime: ...
+#     def start_of_day(self) -> DateTime: ...
+#     def tomorrow(self) -> DateTime: ...
+#     def to_zoned(self, tz: TimeZone) -> ZonedDateTime: ...
+#
+#     # =========================================================================
+#     # SINCE/UNTIL
+#     # =========================================================================
+#     def _since(self, other: DateTimeDifference) -> TimeSpan: ...
+#     def _until(self, other: DateTimeDifference) -> TimeSpan: ...
+#     def since(
+#         self,
+#         other: Date | Time | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def until(
+#         self,
+#         other: Date | Time | DateTime | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#
+# class DateTimeDifference:
+#     def __init__(
+#         self,
+#         date: DateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def smallest(self, unit: JIFF_UNIT_STRING) -> DateTimeDifference: ...
+#     def largest(self, unit: JIFF_UNIT_STRING) -> DateTimeDifference: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateTimeDifference: ...
+#     def increment(self, increment: int) -> DateTimeDifference: ...
+#
+# class TimeZone:
+#     def __init__(self, name: str) -> None: ...
+#     def __eq__(self, other: object) -> bool: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     def to_pytzinfo(self) -> pydt.tzinfo: ...
+#     @classmethod
+#     def from_pytzinfo(cls: type[TimeZone], tz: pydt.tzinfo) -> TimeZone: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def name(self) -> str: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def fixed(cls: type[TimeZone], offset: Offset) -> TimeZone: ...
+#     @classmethod
+#     def get(cls: type[TimeZone], name: str) -> TimeZone: ...
+#     @classmethod
+#     def posix(cls: type[TimeZone], name: str) -> TimeZone: ...
+#     @classmethod
+#     def system(cls: type[TimeZone]) -> TimeZone: ...
+#     @classmethod
+#     def try_system(cls: type[TimeZone]) -> TimeZone: ...
+#     @classmethod
+#     def tzif(cls: type[TimeZone], name: str, data: bytes) -> TimeZone: ...
+#     @classmethod
+#     def utc(cls: type[TimeZone]) -> TimeZone: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def iana_name(self) -> str | None: ...
+#     def to_datetime(self, dt: Timestamp) -> DateTime: ...
+#     def to_offset(self, timestamp: Timestamp) -> tuple[Offset, bool, str]: ...
+#     def to_timestamp(self, dt: DateTime) -> Timestamp: ...
+#     def to_zoned(self, other: DateTime) -> ZonedDateTime: ...
+#
+#     # =========================================================================
+#     # NOT IMPLEMENTED
+#     # =========================================================================
+#     def to_ambiguous_timestamp(self) -> t.NoReturn: ...
+#     def to_ambiguous_zoned(self) -> t.NoReturn: ...
+#
+# class SignedDuration:
+#     MIN: SignedDuration
+#     MAX: SignedDuration
+#     ZERO: SignedDuration
+#
+#     def __init__(self, secs: int = 0, nanos: int = 0) -> None: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#     def __hash__(self) -> int: ...
+#     def __mul__(self, other: int) -> SignedDuration: ...
+#     def __rmul__(self, other: int) -> SignedDuration: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __lt__(self, other: object) -> bool: ...
+#     def __le__(self, other: object) -> bool: ...
+#     def __gt__(self, other: object) -> bool: ...
+#     def __ge__(self, other: object) -> bool: ...
+#     def __neg__(self) -> SignedDuration: ...
+#     def __add__(self, other: SignedDuration) -> SignedDuration: ...
+#     def __abs__(self) -> SignedDuration: ...
+#     def __div__(self, other: int) -> SignedDuration: ...
+#     def abs(self) -> SignedDuration: ...
+#     def unsigned_abs(self) -> Duration: ...
+#     def __richcmp__(self, other: SignedDuration | pydt.timedelta, op: int) -> bool: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def __str__(self) -> str: ...
+#     def string(self, human: bool = False) -> str: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     @classmethod
+#     def from_pytimedelta(
+#         cls: type[SignedDuration], td: pydt.timedelta
+#     ) -> SignedDuration: ...
+#     def to_pytimedelta(self) -> pydt.timedelta: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def parse(cls: type[SignedDuration], s: str) -> SignedDuration: ...
+#     @classmethod
+#     def from_hours(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_micros(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_millis(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_mins(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_nanos(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_secs(cls: type[SignedDuration], n: int) -> SignedDuration: ...
+#     @classmethod
+#     def from_secs_f32(cls: type[SignedDuration], n: float) -> SignedDuration: ...
+#     @classmethod
+#     def from_secs_f64(cls: type[SignedDuration], n: float) -> SignedDuration: ...
+#     @classmethod
+#     def try_from_secs_f32(cls: type[SignedDuration], n: float) -> SignedDuration: ...
+#     @classmethod
+#     def try_from_secs_f64(cls: type[SignedDuration], n: float) -> SignedDuration: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def is_negative(self) -> bool: ...
+#     @property
+#     def is_zero(self) -> bool: ...
+#     @property
+#     def secs(self) -> int: ...
+#     @property
+#     def nanos(self) -> int: ...
+#     @property
+#     def days(self) -> int: ...
+#     @property
+#     def seconds(self) -> int: ...
+#     @property
+#     def microseconds(self) -> int: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def as_hours(self) -> int: ...
+#     def as_micros(self) -> int: ...
+#     def as_millis(self) -> int: ...
+#     def as_millis_f32(self) -> float: ...
+#     def as_millis_f64(self) -> float: ...
+#     def as_mins(self) -> int: ...
+#     def as_nanos(self) -> int: ...
+#     def as_secs(self) -> int: ...
+#     def as_secs_f32(self) -> float: ...
+#     def as_secs_f64(self) -> float: ...
+#     def checked_add(self, other: SignedDuration) -> SignedDuration | None: ...
+#     def checked_div(self, other: int) -> SignedDuration | None: ...
+#     def checked_mul(self, other: int) -> SignedDuration | None: ...
+#     def checked_neg(self) -> SignedDuration | None: ...
+#     def checked_sub(self, other: SignedDuration) -> SignedDuration | None: ...
+#     def div_duration_f32(self, other: SignedDuration) -> float: ...
+#     def div_duration_f64(self, other: SignedDuration) -> float: ...
+#     def div_f32(self, other: int) -> float: ...
+#     def div_f64(self, other: int) -> float: ...
+#     def is_positive(self) -> bool: ...
+#     def mul_f32(self, other: int) -> SignedDuration: ...
+#     def mul_f64(self, other: int) -> SignedDuration: ...
+#     def saturating_add(self, other: SignedDuration) -> SignedDuration: ...
+#     def saturating_mul(self, other: int) -> SignedDuration: ...
+#     def saturating_sub(self, other: SignedDuration) -> SignedDuration: ...
+#     def signum(self) -> t.Literal[-1, 0, 1]: ...
+#     def subsec_micros(self) -> int: ...
+#     def subsec_millis(self) -> int: ...
+#     def subsec_nanos(self) -> int: ...
+#     def to_timespan(self) -> TimeSpan: ...
+#
+# # put in quotes to avoid ruff F821 - undefined name
+# _TimeSpanArithmeticSingle = TimeSpan | Duration | SignedDuration
+# _TimeSpanArithmeticTuple = tuple[
+#     _TimeSpanArithmeticSingle, ZonedDateTime | Date | DateTime
+# ]
+# TimeSpanArithmetic = _TimeSpanArithmeticSingle | _TimeSpanArithmeticTuple
+#
+# class TimeSpan:
+#     def __init__(
+#         self,
+#         years: int = 0,
+#         months: int = 0,
+#         weeks: int = 0,
+#         days: int = 0,
+#         hours: int = 0,
+#         minutes: int = 0,
+#         seconds: int = 0,
+#         milliseconds: int = 0,
+#         microseconds: int = 0,
+#         nanoseconds: int = 0,
+#     ) -> None: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def string(self, human: bool = False) -> str: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def repr_full(self) -> str: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     @classmethod
+#     def from_pytimedelta(cls, td: pydt.timedelta) -> TimeSpan: ...
+#     def to_pytimedelta(self) -> pydt.timedelta: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def parse(cls, s: str) -> TimeSpan: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def is_positive(self) -> bool: ...
+#     @property
+#     def is_negative(self) -> bool: ...
+#     @property
+#     def is_zero(self) -> bool: ...
+#     @property
+#     def years(self) -> int: ...
+#     @property
+#     def months(self) -> int: ...
+#     @property
+#     def weeks(self) -> int: ...
+#     @property
+#     def days(self) -> int: ...
+#     @property
+#     def hours(self) -> int: ...
+#     @property
+#     def minutes(self) -> int: ...
+#     @property
+#     def seconds(self) -> int: ...
+#     @property
+#     def milliseconds(self) -> int: ...
+#     @property
+#     def microseconds(self) -> int: ...
+#     @property
+#     def nanoseconds(self) -> int: ...
+#
+#     # =========================================================================
+#     # OPERATORS
+#     # =========================================================================
+#     def __add__(
+#         self,
+#         val: TimeSpanArithmetic,
+#     ) -> te.Self: ...
+#     def __sub__(
+#         self,
+#         val: TimeSpanArithmetic,
+#     ) -> te.Self: ...
+#     def __mul__(self, other: int) -> te.Self: ...
+#     def __neg__(self) -> te.Self: ...
+#     def __abs__(self) -> te.Self: ...
+#     def __invert__(self) -> te.Self: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ge__(self, other: TimeSpan) -> bool: ...
+#     def __gt__(self, other: TimeSpan) -> bool: ...
+#     def __le__(self, other: TimeSpan) -> bool: ...
+#     def __lt__(self, other: TimeSpan) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __rmul__(self, other: TimeSpan) -> bool: ...
+#     def __hash__(self) -> int: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#
+#     def abs(self) -> te.Self: ...
+#     def asdict(self) -> TimeSpanTypedDict: ...
+#     def checked_add(self, val: TimeSpanArithmetic) -> te.Self: ...
+#     def checked_mul(self, other: int) -> te.Self: ...
+#     def checked_sub(self, val: TimeSpanArithmetic) -> te.Self: ...
+#     def compare(
+#         self, other: TimeSpan, relative: ZonedDateTime | DateTime | Date | None = None
+#     ) -> int: ...
+#     def negate(self) -> te.Self: ...
+#     def replace(
+#         self,
+#         years: int | None = None,
+#         months: int | None = None,
+#         weeks: int | None = None,
+#         days: int | None = None,
+#         hours: int | None = None,
+#         minutes: int | None = None,
+#         seconds: int | None = None,
+#         milliseconds: int | None = None,
+#         microseconds: int | None = None,
+#         nanoseconds: int | None = None,
+#     ) -> te.Self: ...
+#     def round(
+#         self,
+#         smallest: JIFF_UNIT_STRING,
+#         increment: int = 1,
+#         *,
+#         relative: ZonedDateTime | Date | DateTime | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#     ) -> te.Self: ...
+#     def signum(self) -> t.Literal[-1, 0, 1]: ...
+#     def to_signed_duration(
+#         self, relative: ZonedDateTime | Date | DateTime
+#     ) -> SignedDuration: ...
+#     def total(
+#         self,
+#         unit: JIFF_UNIT_STRING,
+#         relative: ZonedDateTime | Date | DateTime | None = None,
+#     ) -> int: ...
+#     def total_seconds(self) -> int: ...
+#     def try_years(self, years: int) -> te.Self: ...
+#     def try_months(self, months: int) -> te.Self: ...
+#     def try_weeks(self, weeks: int) -> te.Self: ...
+#     def try_days(self, days: int) -> te.Self: ...
+#     def try_hours(self, hours: int) -> te.Self: ...
+#     def try_minutes(self, minutes: int) -> te.Self: ...
+#     def try_seconds(self, seconds: int) -> te.Self: ...
+#     def try_milliseconds(self, milliseconds: int) -> te.Self: ...
+#     def try_microseconds(self, microseconds: int) -> te.Self: ...
+#     def try_nanoseconds(self, nanoseconds: int) -> te.Self: ...
+#
+#     # -------------------------------------------------------------------------
+#     # PANIC-INDUCING METHODS
+#     # -------------------------------------------------------------------------
+#     def _years(self, years: int) -> te.Self: ...
+#     def _months(self, months: int) -> te.Self: ...
+#     def _weeks(self, weeks: int) -> te.Self: ...
+#     def _days(self, days: int) -> te.Self: ...
+#     def _hours(self, hours: int) -> te.Self: ...
+#     def _minutes(self, minutes: int) -> te.Self: ...
+#     def _seconds(self, seconds: int) -> te.Self: ...
+#     def _milliseconds(self, milliseconds: int) -> te.Self: ...
+#     def _microseconds(self, microseconds: int) -> te.Self: ...
+#     def _nanoseconds(self, nanoseconds: int) -> te.Self: ...
+#
+# class Timestamp:
+#     """
+#     A representation of a timestamp with second and nanosecond precision.
+#     """
+#
+#     MIN: Timestamp
+#     MAX: Timestamp
+#     UNIX_EPOCH: Timestamp
+#
+#     def __init__(
+#         self, second: int | None = None, nanosecond: int | None = None
+#     ) -> None: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def now(cls) -> Timestamp: ...
+#     @classmethod
+#     def parse(cls, s: str) -> Timestamp: ...
+#     @classmethod
+#     def from_millisecond(cls, millisecond: int) -> Timestamp: ...
+#     @classmethod
+#     def from_microsecond(cls, microsecond: int) -> Timestamp: ...
+#     @classmethod
+#     def from_nanosecond(cls, nanosecond: int) -> Timestamp: ...
+#     @classmethod
+#     def from_second(cls, second: int) -> Timestamp: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#     def __add__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ge__(self, other: Timestamp) -> bool: ...
+#     def __gt__(self, other: Timestamp) -> bool: ...
+#     def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     def __le__(self, other: Timestamp) -> bool: ...
+#     def __lt__(self, other: Timestamp) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __richcmp__(self, other: Timestamp, op: int) -> bool: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS W/ OVERLOADS
+#     # =========================================================================
+#     @t.overload
+#     def __isub__(self, other: Timestamp) -> TimeSpan: ...
+#     @t.overload
+#     def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     @t.overload
+#     def __sub__(self, other: Timestamp) -> TimeSpan: ...
+#     @t.overload
+#     def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#
+#     # =========================================================================
+#     # STRPTIME/STRFTIME
+#     # =========================================================================
+#     def strftime(self, format: str) -> str: ...
+#     @classmethod
+#     def strptime(cls, format: str, input: str) -> Timestamp: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#
+#     def as_microsecond(self) -> int: ...
+#     def as_millisecond(self) -> int: ...
+#     def as_nanosecond(self) -> int: ...
+#     def as_second(self) -> int: ...
+#     def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> Timestamp: ...
+#     @t.overload
+#     def checked_sub(self, other: Timestamp) -> TimeSpan: ...
+#     @t.overload
+#     def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> Timestamp: ...
+#     def display_with_offset(self, offset: Offset) -> str: ...
+#     def in_tz(self, tz: str) -> ZonedDateTime: ...
+#     def intz(self, tz: str) -> ZonedDateTime:
+#         """Deprecated ~ use `in_tz`"""
+#
+#     def is_zero(self) -> bool: ...
+#     def saturating_add(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> Timestamp: ...
+#     def saturating_sub(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> Timestamp: ...
+#     def series(self, span: TimeSpan) -> t.Iterator[Timestamp]: ...
+#     def signum(self) -> t.Literal[-1, 0, 1]: ...
+#     def string(self) -> str: ...
+#     def subsec_microsecond(self) -> int: ...
+#     def subsec_millisecond(self) -> int: ...
+#     def subsec_nanosecond(self) -> int: ...
+#     def to_zoned(self, time_zone: TimeZone) -> ZonedDateTime: ...
+#
+#     # =========================================================================
+#     # SINCE/UNTIL
+#     # =========================================================================
+#     def _since(self, other: TimestampDifference) -> TimeSpan: ...
+#     def _until(self, other: TimestampDifference) -> TimeSpan: ...
+#     def since(
+#         self,
+#         other: Timestamp | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def until(
+#         self,
+#         other: Timestamp | ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def duration_since(self, other: Timestamp) -> SignedDuration: ...
+#     def duration_until(self, other: Timestamp) -> SignedDuration: ...
+#     def round(
+#         self,
+#         unit: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> Timestamp: ...
+#     def _round(self, options: TimestampRound) -> Timestamp: ...
+#
+# class TimestampDifference:
+#     def __init__(
+#         self,
+#         date: Timestamp,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def smallest(self, unit: JIFF_UNIT_STRING) -> TimestampDifference: ...
+#     def largest(self, unit: JIFF_UNIT_STRING) -> TimestampDifference: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimestampDifference: ...
+#     def increment(self, increment: int) -> TimestampDifference: ...
+#
+# class ZonedDateTime:
+#     def __init__(self, timestamp: Timestamp, time_zone: TimeZone) -> None: ...
+#
+#     # =========================================================================
+#     # PYTHON CONVERSIONS
+#     # =========================================================================
+#     @classmethod
+#     def from_pydatetime(
+#         cls: type[ZonedDateTime], dt: pydt.datetime
+#     ) -> ZonedDateTime: ...
+#     def to_pydatetime(self) -> pydt.datetime: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def now(cls: type[ZonedDateTime], tz: str | None = None) -> ZonedDateTime: ...
+#     @classmethod
+#     def utcnow(cls: type[ZonedDateTime]) -> ZonedDateTime: ...
+#     @classmethod
+#     def parse(cls: type[ZonedDateTime], s: str) -> ZonedDateTime: ...
+#     @classmethod
+#     def from_rfc2822(cls: type[ZonedDateTime], s: str) -> ZonedDateTime: ...
+#
+#     # =========================================================================
+#     # STRPTIME/STRFTIME
+#     # =========================================================================
+#     @classmethod
+#     def strptime(
+#         cls: type[ZonedDateTime], format: str, input: str
+#     ) -> ZonedDateTime: ...
+#     def strftime(self, format: str) -> str: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def year(self) -> int: ...
+#     @property
+#     def month(self) -> int: ...
+#     @property
+#     def day(self) -> int: ...
+#     @property
+#     def hour(self) -> int: ...
+#     @property
+#     def minute(self) -> int: ...
+#     @property
+#     def second(self) -> int: ...
+#     @property
+#     def millisecond(self) -> int: ...
+#     @property
+#     def microsecond(self) -> int: ...
+#     @property
+#     def nanosecond(self) -> int: ...
+#     @property
+#     def subsec_nanosecond(self) -> int: ...
+#     @property
+#     def weekday(self) -> int: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def __str__(self) -> str: ...
+#     def string(self) -> str: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#     def __add__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ge__(self, other: ZonedDateTime) -> bool: ...
+#     def __gt__(self, other: ZonedDateTime) -> bool: ...
+#     def __hash__(self) -> int: ...
+#     def __iadd__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     def __le__(self, other: ZonedDateTime) -> bool: ...
+#     def __lt__(self, other: ZonedDateTime) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __richcmp__(self, other: ZonedDateTime, op: int) -> bool: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS W/ OVERLOADS
+#     # =========================================================================
+#     @t.overload
+#     def __isub__(self, other: ZonedDateTime) -> TimeSpan: ...
+#     @t.overload
+#     def __isub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     @t.overload
+#     def __sub__(self, other: ZonedDateTime) -> TimeSpan: ...
+#     @t.overload
+#     def __sub__(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def astimezone(self, tz: str) -> ZonedDateTime: ...
+#     def checked_add(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     @t.overload
+#     def checked_sub(self, other: ZonedDateTime) -> TimeSpan: ...
+#     @t.overload
+#     def checked_sub(self, other: TimeSpan | SignedDuration | Duration) -> te.Self: ...
+#     def date(self) -> Date: ...
+#     def datetime(self) -> DateTime: ...
+#     def iso_week_date(self) -> ISOWeekDate: ...
+#     def day_of_year(self) -> int: ...
+#     def day_of_year_no_leap(self) -> int | None: ...
+#     def days_in_month(self) -> int: ...
+#     def days_in_year(self) -> int: ...
+#     def duration_since(self, other: ZonedDateTime) -> SignedDuration: ...
+#     def duration_until(self, other: ZonedDateTime) -> SignedDuration: ...
+#     def end_of_day(self) -> ZonedDateTime: ...
+#     def era_year(self) -> tuple[int, t.Literal["CE", "BCE"]]: ...
+#     def first_of_month(self) -> ZonedDateTime: ...
+#     def first_of_year(self) -> ZonedDateTime: ...
+#     def in_leap_year(self) -> bool: ...
+#     def in_tz(self, tz: str) -> te.Self: ...
+#     def intz(self, tz: str) -> te.Self: ...
+#     def inutc(self) -> ZonedDateTime: ...
+#     def last_of_month(self) -> ZonedDateTime: ...
+#     def last_of_year(self) -> ZonedDateTime: ...
+#     def nth_weekday(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> Date: ...
+#     def nth_weekday_of_month(
+#         self, nth: int, weekday: JIFF_WEEKDAY_STRING | JIFF_WEEKDAY_INT
+#     ) -> Date: ...
+#     def offset(self) -> Offset: ...
+#     def round(self, options: JIFF_UNIT_STRING | DateTimeRound) -> te.Self: ...
+#     def saturating_add(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> te.Self: ...
+#     @t.overload
+#     def saturating_sub(self, other: ZonedDateTime) -> TimeSpan: ...
+#     @t.overload
+#     def saturating_sub(
+#         self, other: TimeSpan | SignedDuration | Duration
+#     ) -> te.Self: ...
+#     def start_of_day(self) -> ZonedDateTime: ...
+#     def time(self) -> Time: ...
+#     def time_zone(self) -> TimeZone: ...
+#     def timestamp(self) -> Timestamp: ...
+#     def timezone(self) -> TimeZone: ...
+#     def to_rfc2822(self) -> str: ...
+#     def tomorrow(self) -> ZonedDateTime: ...
+#     def with_time_zone(self, tz: TimeZone) -> ZonedDateTime: ...
+#     def yesterday(self) -> ZonedDateTime: ...
+#
+#     # =========================================================================
+#     # SINCE/UNTIL
+#     # =========================================================================
+#     def since(
+#         self,
+#         other: ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#     def until(
+#         self,
+#         other: ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> TimeSpan: ...
+#
+# class ZonedDateTimeDifference:
+#     def __init__(
+#         self,
+#         date: ZonedDateTime,
+#         *,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         largest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int | None = None,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def smallest(self, unit: JIFF_UNIT_STRING) -> ZonedDateTimeDifference: ...
+#     def largest(self, unit: JIFF_UNIT_STRING) -> ZonedDateTimeDifference: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> ZonedDateTimeDifference: ...
+#     def increment(self, increment: int) -> ZonedDateTimeDifference: ...
+#
+# class ISOWeekDate:
+#     MIN: ISOWeekDate
+#     MAX: ISOWeekDate
+#     ZERO: ISOWeekDate
+#
+#     def __init__(
+#         self, year: int, week: int, weekday: JIFF_WEEKDAY_INT | JIFF_WEEKDAY_STRING
+#     ) -> None: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __lt__(self, other: ISOWeekDate) -> bool: ...
+#     def __le__(self, other: ISOWeekDate) -> bool: ...
+#     def __gt__(self, other: ISOWeekDate) -> bool: ...
+#     def __ge__(self, other: ISOWeekDate) -> bool: ...
+#     def __hash__(self) -> int: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # CLASS METHODS
+#     # =========================================================================
+#     @classmethod
+#     def from_date(cls: type[ISOWeekDate], date: Date) -> ISOWeekDate: ...
+#     @classmethod
+#     def today(cls: type[ISOWeekDate]) -> ISOWeekDate: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def year(self) -> int: ...
+#     @property
+#     def week(self) -> int: ...
+#     @property
+#     def weekday(self) -> JIFF_WEEKDAY_INT: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def date(self) -> Date: ...
+#
+# class TimestampRound:
+#     def __init__(
+#         self,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int = 1,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> TimestampRound: ...
+#     def smallest(self, smallest: JIFF_UNIT_STRING) -> TimestampRound: ...
+#     def increment(self, increment: int) -> TimestampRound: ...
+#     def _smallest(self) -> JIFF_UNIT_STRING: ...
+#     def _mode(self) -> JIFF_ROUND_MODE_STRING: ...
+#     def _increment(self) -> int: ...
+#     def replace(
+#         self,
+#         smallest: JIFF_UNIT_STRING | None,
+#         mode: JIFF_ROUND_MODE_STRING | None,
+#         increment: int | None,
+#     ) -> TimestampRound: ...
+#
+# class DateTimeRound:
+#     def __init__(
+#         self,
+#         smallest: JIFF_UNIT_STRING | None = None,
+#         mode: JIFF_ROUND_MODE_STRING | None = None,
+#         increment: int = 1,
+#     ) -> None: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def mode(self, mode: JIFF_ROUND_MODE_STRING) -> DateTimeRound: ...
+#     def smallest(self, smallest: JIFF_UNIT_STRING) -> DateTimeRound: ...
+#     def increment(self, increment: int) -> DateTimeRound: ...
+#     def _smallest(self) -> JIFF_UNIT_STRING: ...
+#     def _mode(self) -> JIFF_ROUND_MODE_STRING: ...
+#     def _increment(self) -> int: ...
+#     def replace(
+#         self,
+#         smallest: JIFF_UNIT_STRING | None,
+#         mode: JIFF_ROUND_MODE_STRING | None,
+#         increment: int | None,
+#     ) -> DateTimeRound: ...
+#
+# class Offset:
+#     MIN: Offset
+#     MAX: Offset
+#     UTC: Offset
+#     ZERO: Offset
+#
+#     def __init__(
+#         self, hours: int | None = None, seconds: int | None = None
+#     ) -> None: ...
+#
+#     # =========================================================================
+#     # STRING
+#     # =========================================================================
+#     def string(self) -> str: ...
+#     def __str__(self) -> str: ...
+#     def __repr__(self) -> str: ...
+#
+#     # =========================================================================
+#     # OPERATORS/DUNDERS
+#     # =========================================================================
+#     def __neg__(self) -> Offset: ...
+#     def __eq__(self, other: object) -> bool: ...
+#     def __ne__(self, other: object) -> bool: ...
+#     def __lt__(self, other: Offset) -> bool: ...
+#     def __le__(self, other: Offset) -> bool: ...
+#     def __gt__(self, other: Offset) -> bool: ...
+#     def __ge__(self, other: Offset) -> bool: ...
+#     def __hash__(self) -> int: ...
+#
+#     # =========================================================================
+#     # PROPERTIES
+#     # =========================================================================
+#     @property
+#     def seconds(self) -> int: ...
+#     @property
+#     def is_negative(self) -> bool: ...
+#     @property
+#     def is_positive(self) -> bool: ...
+#
+#     # =========================================================================
+#     # FROM
+#     # =========================================================================
+#     @classmethod
+#     def utc(cls: type[Offset]) -> Offset: ...
+#     @classmethod
+#     def from_hours(cls: type[Offset], hours: int) -> Offset: ...
+#     @classmethod
+#     def from_seconds(cls: type[Offset], seconds: int) -> Offset: ...
+#
+#     # =========================================================================
+#     # TO
+#     # =========================================================================
+#     def to_datetime(self, timestamp: Timestamp) -> DateTime: ...
+#     def to_timestamp(self, datetime: DateTime) -> Timestamp: ...
+#     def to_timezone(self) -> TimeZone: ...
+#
+#     # =========================================================================
+#     # INSTANCE METHODS
+#     # =========================================================================
+#     def checked_add(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
+#     def checked_sub(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
+#     def duration_since(self, other: Offset) -> SignedDuration: ...
+#     def duration_until(self, other: Offset) -> SignedDuration: ...
+#     def negate(self) -> Offset: ...
+#     def saturating_add(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
+#     def saturating_sub(self, other: Duration | SignedDuration | TimeSpan) -> Offset: ...
+#     def since(self, other: Offset) -> TimeSpan: ...
+#     def until(self, other: Offset) -> TimeSpan: ...
+#
+# def date(year: int, month: int, day: int) -> Date: ...
+# def time(
+#     hour: int = 0, minute: int = 0, second: int = 0, nanosecond: int = 0
+# ) -> Time: ...
+# def datetime(
+#     year: int,
+#     month: int,
+#     day: int,
+#     hour: int = 0,
+#     minute: int = 0,
+#     second: int = 0,
+#     nanosecond: int = 0,
+# ) -> DateTime: ...
+# def timespan(
+#     *,
+#     years: int = 0,
+#     months: int = 0,
+#     weeks: int = 0,
+#     days: int = 0,
+#     hours: int = 0,
+#     minutes: int = 0,
+#     seconds: int = 0,
+#     milliseconds: int = 0,
+#     microseconds: int = 0,
+#     nanoseconds: int = 0,
+#     unchecked: bool = False,
+# ) -> TimeSpan: ...
+# def offset(hours: int) -> Offset: ...
